Fixes #45
