--- backend/computer-backend.php
@@ -1,13 +1,0 @@
-<?php 
-	$postType = "computer";
-    $page = "Computer";
-
-    require_once('header-backend.php');
-    require_once('./../config/dbconfig.php');
-    require_once('./../functions.php');
-   
-    /* display page content */ 
-    require_once('template-backend.php'); 
-?>
-
-<?php require_once('footer-backend.php') ?>
