+++ backend/delete-backend.php
@@ -9,10 +9,10 @@
 		$result = $conn->query($sql);
 
 		if ($result) {
-			header("location: news-backend.php?status=success");
+			header('location: backend.php?status=success&postType='.$postType.'&page='.$page);
 		}
 		else {
-			header("location: news-backend.php?status=fail");
+			header('location: news-backend.php?status=fail&postType='.$postType.'&page='.$page);
 		}
 	}
 ?>
