+++ backend/list-backend.php
@@ -1,6 +1,6 @@
 <?php 
 	require_once('./../config/dbconfig.php');
-	require_once('./../functions.php');
+	require_once('functions-backend.php');
 ?>
 
 <?php
