+++ backend/navbar-backend.php
@@ -1,4 +1,5 @@
 <?php 
+    $dashboard = '';
     $news = '';
     $computer = '';
     $phone = '';
@@ -22,7 +23,15 @@
         case 'wiki':
             $wiki = 'class="active"';
         break;
+        default:
+            $dashboard = 'class="active"';
     }
+
+    $urlNews = 'backend.php?postType=news&page=News';
+    $urlComputer = 'backend.php?postType=computer&page=Computer';
+    $urlPhone = 'backend.php?postType=phone&page=Phone';
+    $urlCamera = 'backend.php?postType=Camera&page=Camera';
+    $urlWiki = 'backend.php?postType=wiki&page=Wiki';
 ?>
 
 <!-- main-navbar -->
@@ -32,27 +41,27 @@
             <h2>TechnoTech</h2>
         </div>
         <div id="menu-backend">
-            <a href="index-backend.php" title="dashboard">
+            <a href="dashboard.php" title="dashboard" <?php echo $dashboard ?>>
                 <span class="fa fa-dashboard"></span>
                 <span class="sub-menu">Dashboard</span>
             </a>
-            <a href="news-backend.php" title="news" <?php echo $news ?>>
+            <a href="<?php echo $urlNews ?>" title="news" <?php echo $news ?>>
                 <span class="fa fa-newspaper-o"></span>
                 <span class="sub-menu">News</span>
             </a>
-            <a href="computer-backend.php" title="computer" <?php echo $computer ?>>
+            <a href="<?php echo $urlComputer ?>" title="computer" <?php echo $computer ?>>
                 <span class="fa fa-desktop"></span>
                 <span class="sub-menu">Computer</span>
             </a>
-            <a href="phone-backend.php" title="phone" <?php echo $phone ?>>
+            <a href="<?php echo $urlPhone ?>" title="phone" <?php echo $phone ?>>
                 <span class="fa fa-mobile"></span>
                 <span class="sub-menu">Phone</span>
             </a>
-            <a href="camera-backend.php" title="camera" <?php echo $camera ?>>
+            <a href="<?php echo $urlCamera ?>" title="camera" <?php echo $camera ?>>
                 <span class="fa fa-camera"></span>
                 <span class="sub-menu">Camera</span>
             </a>
-            <a href="wiki-backend.php" title="Wiki" <?php echo $wiki ?>>
+            <a href="<?php echo $urlWiki ?>" title="Wiki" <?php echo $wiki ?>>
                 <span class="fa fa-book"></span>
                 <span class="sub-menu">Wiki</span>
             </a>
